ARM: 8160/1: drop warning about return_address not using unwind tables
The warning was introduced in 2009 (commit 4bf1fa5 ([ARM] 5613/1:
implement CALLER_ADDRESSx)). The only 'problem' here is that
CALLER_ADDRESSx for x > 1 returns NULL which doesn't do much harm.

The drawback of implementing a fix
(i.e. use unwind tables to implement CALLER_ADDRESSx) is that much of the
unwinder code would need to be marked as not traceable.

Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>